Login Feature Corrected
Fixed sign up error where database kept stating that the email column did not exist. As a result, a modification was made to the app.py to dynamically resolve the issue. If not stated previously, additional information is asked and collected from the user such as email, name, age, etc. Terms and conditions was also added as a temporary placeholder.